Don't rebuild cluster
We were deleting existing VM and rebuilding it if the cluster didn't
have VM and cluster running, which could be problematic.

This patch changes this behavior and fails the run with an error message
if that happens so we can manually explore what's going on or delete the
VM ourselves.