Autoselect iscsid tag
The script only supported the "latest" tag for the iSCSI container
daemon (which is CentOS 8 based), which meant that the script wouldn't
work out of the box without a configuration file, since Ember-CSI's
master container uses a CentOS 7 base.

This patch moves the step where we deploy the iSCSI daemon to after the
driver has been deployed, allowing us to detect the OS version of the
Ember-CSI container and deploy the right iSCSI daemon.