Build with host cache
Latest Dockerfile in the Ember-CSI project is a multistage build that
requires the system cache to be mounted on the host to make the packages
generated in the build stage available for the final container.

This patch introduces the cache feature and will always build containers
using it.

New configuration option called CACHE_DIR