Add CSC command
Adds the possibility of running CSI commands directly against the
controller or node using the csc command.

This is useful after running csi-sanity tests to do cleanups, because
volumes and snapshots created by the tests won't have corresponding
Kubernetes/OpenShift resources we can use to operate on the resources.

Example:
  ./start.sh csc config -e /controller.sock controller list-volumes